genAI/ai_meme_generator/requirements.txt
jupyter
ipywidgets
google-generativeai
Pillow
fastapi
uvicorn
python-multipart
